+++ Sources/Plasma/NucleusLib/pnModifier/plModifier.h
@@ -65,7 +65,7 @@ protected:
     plCoordinateInterface*      IGetTargetCoordinateInterface(int iTarg) const;
     plAudioInterface*           IGetTargetAudioInterface(int iTarg) const;
     plObjInterface*             IGetTargetGenericInterface(int iTarg, uint32_t classIdx) const;
-    plModifier*                 IGetTargetModifier(int iTarg, int iMod) const;
+    plModifier*                 IGetTargetModifier(int iTarg, size_t iMod) const;
 
     virtual bool IEval(double secs, float del, uint32_t dirty) = 0; // called only by owner object's Eval()
