+++ Sources/Plasma/PubUtilLib/plModifier/plAGMasterSDLModifier.cpp
@@ -122,9 +122,9 @@ void plAGMasterSDLModifier::IPutCurrentStateIn(plStateDataRecord* dstState)
 //
 plAGMasterMod* plAGMasterSDLModifier::IGetObjectsAGMasterMod(plSceneObject* obj)
 {
-    int count = obj->GetNumModifiers();
+    size_t count = obj->GetNumModifiers();
 
-    for (int i = 0; i < count; i++)
+    for (size_t i = 0; i < count; i++)
     {
         plAGMasterMod * avMod = const_cast<plAGMasterMod*>(plAGMasterMod::ConvertNoRef(obj->GetModifier(i)));
         if(avMod)
